# swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "My Nigerian Contacts & Projects API",
    "description": "Contacts and Projects API for CSE 341 - Performs CRUD operations on MongoDB"
  },
  "host": "cse341-contacts-unhb.onrender.com",
  "basePath": "/",
  "schemes": [
    "https",
    "http"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/api-docs": {
      "get": {
        "description": "API Documentation",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/contacts": {
      "get": {
        "tags": [
          "Contacts"
        ],
        "description": "Get all contacts",
        "responses": {
          "200": {
            "description": "Successfully retrieved all contacts",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Contact"
              }
            }
          },
          "500": {
            "description": "Server error"
          }
        }
      },
      "post": {
        "tags": [
          "Contacts"
        ],
        "description": "Create a new contact",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Contact_Input"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Contact created successfully",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string"
                },
                "insertedId": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Validation error"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
    "/contacts/{id}": {
      "get": {
        "tags": [
          "Contacts"
        ],
        "description": "Get a single contact by ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "Contact MongoDB ObjectId"
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved contact",
            "schema": {
              "$ref": "#/definitions/Contact"
            }
          },
          "400": {
            "description": "Invalid ID format"
          },
          "404": {
            "description": "Contact not found"
          },
          "500": {
            "description": "Server error"
          }
        }
      },
      "put": {
        "tags": [
          "Contacts"
        ],
        "description": "Update a contact",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "Contact MongoDB ObjectId"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Contact_Input"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Contact updated successfully"
          },
          "400": {
            "description": "Validation error"
          },
          "404": {
            "description": "Contact not found"
          },
          "500": {
            "description": "Server error"
          }
        }
      },
      "delete": {
        "tags": [
          "Contacts"
        ],
        "description": "Delete a contact",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "Contact MongoDB ObjectId"
          }
        ],
        "responses": {
          "200": {
            "description": "Contact deleted successfully"
          },
          "400": {
            "description": "Invalid ID format"
          },
          "404": {
            "description": "Contact not found"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
    "/projects": {
      "get": {
        "tags": [
          "Projects"
        ],
        "description": "Get all projects",
        "responses": {
          "200": {
            "description": "Successfully retrieved all projects",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Project"
              }
            }
          },
          "500": {
            "description": "Server error"
          }
        }
      },
      "post": {
        "tags": [
          "Projects"
        ],
        "description": "Create a new project",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Project_Input"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Project created successfully",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string"
                },
                "insertedId": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Validation error"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
    "/projects/{id}": {
      "get": {
        "tags": [
          "Projects"
        ],
        "description": "Get a single project by ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "Project MongoDB ObjectId"
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved project",
            "schema": {
              "$ref": "#/definitions/Project"
            }
          },
          "400": {
            "description": "Invalid ID format"
          },
          "404": {
            "description": "Project not found"
          },
          "500": {
            "description": "Server error"
          }
        }
      },
      "put": {
        "tags": [
          "Projects"
        ],
        "description": "Update a project",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "Project MongoDB ObjectId"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Project_Input"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Project updated successfully"
          },
          "400": {
            "description": "Validation error"
          },
          "404": {
            "description": "Project not found"
          },
          "500": {
            "description": "Server error"
          }
        }
      },
      "delete": {
        "tags": [
          "Projects"
        ],
        "description": "Delete a project",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "Project MongoDB ObjectId"
          }
        ],
        "responses": {
          "200": {
            "description": "Project deleted successfully"
          },
          "400": {
            "description": "Invalid ID format"
          },
          "404": {
            "description": "Project not found"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    }
  },
  "definitions": {
    "Contact": {
      "type": "object",
      "properties": {
        "_id": {
          "type": "string",
          "description": "MongoDB ObjectId"
        },
        "firstName": {
          "type": "string",
          "example": "John"
        },
        "lastName": {
          "type": "string",
          "example": "Doe"
        },
        "email": {
          "type": "string",
          "example": "john@example.com"
        },
        "phone": {
          "type": "string",
          "example": "+234812345678"
        },
        "address": {
          "type": "string",
          "example": "123 Main St, Lagos"
        },
        "profession": {
          "type": "string",
          "example": "Software Engineer"
        },
        "favoriteColor": {
          "type": "string",
          "example": "Blue"
        },
        "birthday": {
          "type": "string",
          "example": "1990-01-15"
        }
      },
      "required": [
        "firstName",
        "lastName",
        "email"
      ]
    },
    "Contact_Input": {
      "type": "object",
      "properties": {
        "firstName": {
          "type": "string",
          "example": "John"
        },
        "lastName": {
          "type": "string",
          "example": "Doe"
        },
        "email": {
          "type": "string",
          "example": "john@example.com"
        },
        "phone": {
          "type": "string",
          "example": "+234812345678"
        },
        "address": {
          "type": "string",
          "example": "123 Main St, Lagos"
        },
        "profession": {
          "type": "string",
          "example": "Software Engineer"
        },
        "favoriteColor": {
          "type": "string",
          "example": "Blue"
        },
        "birthday": {
          "type": "string",
          "example": "1990-01-15"
        }
      },
      "required": [
        "firstName",
        "lastName",
        "email"
      ]
    },
    "Project": {
      "type": "object",
      "properties": {
        "_id": {
          "type": "string",
          "description": "MongoDB ObjectId"
        },
        "name": {
          "type": "string",
          "example": "Website Redesign"
        },
        "description": {
          "type": "string",
          "example": "Redesign company website with modern UI"
        },
        "status": {
          "type": "string",
          "enum": [
            "pending",
            "in_progress",
            "completed",
            "on_hold"
          ],
          "example": "in_progress"
        },
        "startDate": {
          "type": "string",
          "example": "2024-01-15"
        },
        "endDate": {
          "type": "string",
          "example": "2024-03-15"
        },
        "priority": {
          "type": "string",
          "enum": [
            "low",
            "medium",
            "high"
          ],
          "example": "high"
        },
        "tags": {
          "type": "array",
          "items": {
            "type": "string"
          },
          "example": [
            "frontend",
            "design"
          ]
        },
        "createdAt": {
          "type": "string",
          "format": "date-time"
        },
        "updatedAt": {
          "type": "string",
          "format": "date-time"
        }
      },
      "required": [
        "name",
        "description",
        "status"
      ]
    },
    "Project_Input": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string",
          "example": "Website Redesign"
        },
        "description": {
          "type": "string",
          "example": "Redesign company website with modern UI"
        },
        "status": {
          "type": "string",
          "enum": [
            "pending",
            "in_progress",
            "completed",
            "on_hold"
          ],
          "example": "in_progress"
        },
        "startDate": {
          "type": "string",
          "example": "2024-01-15"
        },
        "endDate": {
          "type": "string",
          "example": "2024-03-15"
        },
        "priority": {
          "type": "string",
          "enum": [
            "low",
            "medium",
            "high"
          ],
          "example": "high"
        },
        "tags": {
          "type": "array",
          "items": {
            "type": "string"
          },
          "example": [
            "frontend",
            "design"
          ]
        }
      },
      "required": [
        "name",
        "description",
        "status"
      ]
    }
  }
}